feat: Add ListConnectedUsers component
The ListConnectedUsers component is added to the client UI to display a list of connected users. This component takes an array of UserConnectedEntity objects as input and renders a ConnectedUserRow component for each user, showing their username. This change improves the user interface by providing a visual representation of the connected users.