feat: Add VoiceChannelProps component
The VoiceChannelProps component is added to the client feature in order to display the voice channel properties. This component takes a ChannelEntity object and an array of UserConnectedEntity objects as input and renders the DisplayChannel component to show the channel details, and the ListConnectedUsers component to display the list of connected users. This change enhances the user interface by providing a visual representation of the voice channel properties.