feat: Add ListChannels component
The ListChannels component is added to the client UI to display a list of voice channels. This component takes an array of ChannelEntity objects and an array of OccupiedChannelEntity objects as input. It renders the VoiceChannelProps component for each channel, showing the channel details and the list of connected users. This change enhances the user interface by providing a visual representation of the voice channels.